requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

numpy==1.22.4
numba
scipy
Unidecode
torch==1.13.1
pypinyin
jieba
pypinyin_dict
